+++ src/geefetch/data/satellites/__init__.py
@@ -2,6 +2,7 @@ from .abc import SatelliteABC
 from .dynworld import DynWorld
 from .gedi import GEDIraster, GEDIvector
 from .landsat8 import Landsat8
+from .palsar2 import Palsar2
 from .s1 import S1
 from .s2 import S2
 
@@ -13,4 +14,5 @@ __all__ = [
     "DynWorld",
     "GEDIraster",
     "Landsat8",
+    "Palsar2",
 ]